chore(package.json): add localstorage-polyfill package to dependencies
fix(server.ts): import order in server.ts to improve readability and maintain consistency

feat(server.ts): change port variable case from lowercase port to uppercase PORT to improve semantics
feat(server.ts): add support for process.env.PORT environment variable to be able to run app on a configurable port

fix(app.component.ts): remove unused imports and add missing imports to improve code quality and maintainability
feat(app.component.ts): add logic to initialize isDark state from local storage and update it on changes

fix(definition-card.component.ts): remove unused imports to improve code quality and maintainability

fix(footer.component.ts): remove unused imports to improve code quality and maintainability

fix(searchbar.component.ts): remove unused imports to improve code quality and maintainability

fix(main-nav.component.ts): add missing class to header element to improve styling

fix(theme-switch.component.ts): add missing imports and implement OnInit interface to improve code quality and maintainability
feat(theme-switch.component.ts): add logic to initialize isChecked state from local storage and update it on changes

chore(word-display.component.ts): update template and remove unused imports for word-display component
feat(shared.service.ts): add dependency injection for StorageService in SharedService constructor
feat(storage.service.ts): add platformId injection in StorageService constructor and implement isLocalStorageAvailable method to check if local storage is available
style(_base.scss): update container margins and add media queries for different screen sizes

+++ src/app/storage.service.ts
@@ -1,25 +1,46 @@
-import { Injectable } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
+import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
 
 @Injectable({
   providedIn: 'root',
 })
 export class StorageService {
-  constructor() {}
+  constructor(@Inject(PLATFORM_ID) private platformId: Object) {}
+
+  isLocalStorageAvailable(): boolean {
+    return isPlatformBrowser(this.platformId) && !!window.localStorage;
+  }
+
+  // isLocalStorageAvailable(): boolean {
+  //   try {
+  //     return typeof window !== 'undefined' && !!window.localStorage;
+  //   } catch (e) {
+  //     return false;
+  //   }
+  // }
 
   setItem(key: string, value: any): void {
-    localStorage.setItem(key, JSON.stringify(value));
+    if (this.isLocalStorageAvailable()) {
+      localStorage.setItem(key, JSON.stringify(value));
+    }
   }
 
   getItem(key: string): any {
-    const item = localStorage.getItem(key);
-    return item ? JSON.parse(item) : null;
+    if (this.isLocalStorageAvailable()) {
+      const item = localStorage.getItem(key);
+      return item ? JSON.parse(item) : null;
+    }
   }
 
   removeItem(key: string): void {
-    localStorage.removeItem(key);
+    if (this.isLocalStorageAvailable()) {
+      localStorage.removeItem(key);
+    }
   }
 
   clear(): void {
-    localStorage.clear();
+    if (this.isLocalStorageAvailable()) {
+      localStorage.clear();
+    }
   }
 }
